Refactor ChestValue to be in inventory
